+++ examples/data-sources/fmc_fqdn_objects/data-source.tf
@@ -1,6 +1,0 @@
-data "fmc_fqdn_objects" "example" {
-  items = {
-    "my_fqdn_objects" = {
-    }
-  }
-}