+++ examples/resources/fmc_fqdns/import.sh
@@ -1,2 +1,3 @@
 # <domain> is optional. If not provided, `Global` is used implicitly and resource's `domain` attribute is not set.
-terraform import fmc_fqdn_object.example "<domain>,<id>"
+terraform import fmc_fqdns.example "<domain>,[<item1_name>,<item2_name>,...]"
+