+++ examples/resources/fmc_fqdns/resource.tf
@@ -1,6 +1,6 @@
-resource "fmc_fqdn_objects" "example" {
+resource "fmc_fqdns" "example" {
   items = {
-    my_fqdn_objects = {
+    my_fqdns = {
       description    = "My FQDN 1"
       fqdn           = "www.example.com"
       dns_resolution = "IPV4_AND_IPV6"