+++ internal/provider/data_source_fmc_fqdns_test.go
@@ -28,21 +28,21 @@ import (
 
 // Section below is generated&owned by "gen/generator.go". //template:begin testAccDataSource
 
-func TestAccDataSourceFmcFQDNObjects(t *testing.T) {
+func TestAccDataSourceFmcFQDNs(t *testing.T) {
 	var checks []resource.TestCheckFunc
-	checks = append(checks, resource.TestCheckResourceAttrSet("data.fmc_fqdn_objects.test", "items.my_fqdn_objects.id"))
-	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_fqdn_objects.test", "items.my_fqdn_objects.description", "My FQDN 1"))
-	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_fqdn_objects.test", "items.my_fqdn_objects.overridable", "true"))
-	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_fqdn_objects.test", "items.my_fqdn_objects.fqdn", "www.example.com"))
-	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_fqdn_objects.test", "items.my_fqdn_objects.dns_resolution", "IPV4_AND_IPV6"))
-	checks = append(checks, resource.TestCheckResourceAttrSet("data.fmc_fqdn_objects.test", "items.my_fqdn_objects.type"))
+	checks = append(checks, resource.TestCheckResourceAttrSet("data.fmc_fqdns.test", "items.my_fqdns.id"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_fqdns.test", "items.my_fqdns.description", "My FQDN 1"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_fqdns.test", "items.my_fqdns.overridable", "true"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_fqdns.test", "items.my_fqdns.fqdn", "www.example.com"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.fmc_fqdns.test", "items.my_fqdns.dns_resolution", "IPV4_AND_IPV6"))
+	checks = append(checks, resource.TestCheckResourceAttrSet("data.fmc_fqdns.test", "items.my_fqdns.type"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 		ErrorCheck:               func(err error) error { return testAccErrorCheck(t, err) },
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDataSourceFmcFQDNObjectsConfig(),
+				Config: testAccDataSourceFmcFQDNsConfig(),
 				Check:  resource.ComposeTestCheckFunc(checks...),
 			},
 		},
@@ -56,9 +56,9 @@ func TestAccDataSourceFmcFQDNObjects(t *testing.T) {
 
 // Section below is generated&owned by "gen/generator.go". //template:begin testAccDataSourceConfig
 
-func testAccDataSourceFmcFQDNObjectsConfig() string {
-	config := `resource "fmc_fqdn_objects" "test" {` + "\n"
-	config += `	items = { "my_fqdn_objects" = {` + "\n"
+func testAccDataSourceFmcFQDNsConfig() string {
+	config := `resource "fmc_fqdns" "test" {` + "\n"
+	config += `	items = { "my_fqdns" = {` + "\n"
 	config += `		description = "My FQDN 1"` + "\n"
 	config += `		overridable = true` + "\n"
 	config += `		fqdn = "www.example.com"` + "\n"
@@ -67,10 +67,10 @@ func testAccDataSourceFmcFQDNObjectsConfig() string {
 	config += `}` + "\n"
 
 	config += `
-		data "fmc_fqdn_objects" "test" {
-			depends_on = [fmc_fqdn_objects.test]
+		data "fmc_fqdns" "test" {
+			depends_on = [fmc_fqdns.test]
 			items = {
-				"my_fqdn_objects" = {
+				"my_fqdns" = {
 				}
 			}
 		}