+++ internal/provider/resource_fmc_host_test.go
@@ -33,7 +33,7 @@ func TestAccFmcHost(t *testing.T) {
 	var checks []resource.TestCheckFunc
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_host.test", "name", "my_host"))
 	checks = append(checks, resource.TestCheckResourceAttrSet("fmc_host.test", "type"))
-	checks = append(checks, resource.TestCheckResourceAttr("fmc_host.test", "description", "My host object"))
+	checks = append(checks, resource.TestCheckResourceAttr("fmc_host.test", "description", "My Host object"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_host.test", "ip", "10.1.1.1"))
 	checks = append(checks, resource.TestCheckResourceAttr("fmc_host.test", "overridable", "true"))
 
@@ -82,7 +82,7 @@ func testAccFmcHostConfig_minimum() string {
 func testAccFmcHostConfig_all() string {
 	config := `resource "fmc_host" "test" {` + "\n"
 	config += `	name = "my_host"` + "\n"
-	config += `	description = "My host object"` + "\n"
+	config += `	description = "My Host object"` + "\n"
 	config += `	ip = "10.1.1.1"` + "\n"
 	config += `	overridable = true` + "\n"
 	config += `}` + "\n"